+++ src/main/java/module-info.java
@@ -1,0 +1,3 @@
+module checksum.up {
+	requires crypt.api.main;
+}